requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.4.0
pillow==11.2.1
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
